requirements.txt
Flask>=2.0
cryptography>=40.0
python-dotenv>=1.0
